Have SWFStream::read_d64() use the convert_double_wacky function to correctly return a double.

+++ libcore/SWFStream.h
@@ -126,7 +126,7 @@ public:
 	//
 	/// aligned read
 	///
-	long double read_d64();
+	double read_d64();
 
 	/// Consume all bits of current byte
 	//
